# Add (and use) bootsnap (for boot time speedup)