--- a/Gemfile
+++ b/Gemfile
@@ -11,6 +11,7 @@ end
 gem 'active_model_serializers'
 gem 'administrate'
 gem 'aws-sdk-s3', require: false
+gem 'bootsnap', require: false
 gem 'browser'
 gem 'connection_pool'
 gem 'dalli'