# Performance testing

_Spec: spec/lib/email/mailgun_via_http_spec.rb_

~Time for 30 runs with `spring`, without `bootsnap`: 72 seconds ("baseline" time)

~Time for 15 runs without `spring`, without `bootsnap`: 190 seconds
Estimated time for 30 runs without `spring`: 380 seconds (5.28x as long as baseline)

~Time for 10 runs without `spring`, with `bootsnap`: 71 seconds
Estimated time for 30 runs without `spring`, with `bootsnap`: 213 seconds (2.96x as long as baseline)

~Time for 30 runs with `spring` _and_ `bootsnap`: 58 seconds (0.81x as long as baseline)

---------------

Time to boot `DISABLE_SPRING=1 bin/rails console` without bootsnap: ~12 seconds

Time to boot `DISABLE_SPRING=1 bin/rails console` with bootsnap: ~6 seconds